+++ Code/ww3d2/dazzle.cpp
@@ -65,6 +65,7 @@
 #include "wwprofile.h"
 #include "visrasterizer.h"
 #include <cstdio>
+#include <stdlib.h>
 #include <limits.h>
 #include <wwprofile.h>
 
@@ -954,7 +955,7 @@ void DazzleRenderObjClass::Render(RenderInfoClass & rinfo)
 
 			unsigned time_ms=WW3D::Get_Frame_Time();
 			if (time_ms==0) time_ms=1;
-			float weight=pow(params->ic.history_weight,time_ms);
+			float weight=pow(params->ic.history_weight,(float)time_ms);
 
 			if (dazzle_intensity>0.0f) {
 				visibility = _VisibilityHandler->Compute_Dazzle_Visibility(rinfo,this,loc);
